scan.cc&friends: Increase variable sizes from unsigned int to size_t to allow for more than 179 mill points per scan on 64 bit architectures
The unsigned int type is 4 byte wide on all platforms. This means that it can
only store up to about 4.2 billion. This means, that when using this type to
address memory made of 3-tuples of doubles (which are 8 byte on all platforms),
then we are down to only 179 million possible 3-tuples because 4.2 billion / 8
/ 3 == 179 million.

Prior to this commit, if any scan exceeded 179 million points, a segmentation
fault would occur.  Thus, this commit changes all relevant size types from
unsigned int to size_t which is a 8 byte wide type on 64 bit platforms. This
allows up to 18 quintillion points per scan on architectures with an 8 byte
wide size_t.

To give a reasonable error message (instead of just a segmentation fault) on
limited 32bit architectures, this commit also adds a check before any
multiplication of point vector sizes to make sure that the result is still
within the bounds of size_t. This check is only be done on platforms where
size_t is 4 bytes long because we assume that nobody will ever need more than
18 quintillion points per scan (famous last words). Doing this check will not
add significant overhead on 32 bit platforms because it is not done in any
inner loop. On 64bit platforms this check will not impact runtime at all
because the if-block will be optimized away by the compiler with optimizations
are switched on.

While we are at it we also fix the existing unsigned int limit for frames and
bump that to size_t as well.

git-svn-id: https://svn.code.sf.net/p/slam6d/code/trunk@1095 51f1e6e7-7289-49ad-8748-f09e43fc6daa

+++ include/slam6d/basicScan.h
@@ -31,14 +31,14 @@ public:
 
   virtual DataPointer get(const std::string& identifier);
   virtual void get(unsigned int types);
-  virtual DataPointer create(const std::string& identifier, unsigned int size);
+  virtual DataPointer create(const std::string& identifier, size_t size);
   virtual void clear(const std::string& identifier);
-  virtual unsigned int readFrames();
+  virtual size_t readFrames();
   virtual void saveFrames();
   virtual void saveBOctTree(std::string & filename);
 
-  virtual unsigned int getFrameCount();
-  virtual void getFrame(unsigned int i, const double*& pose_matrix, AlgoType& type);
+  virtual size_t getFrameCount();
+  virtual void getFrame(size_t i, const double*& pose_matrix, AlgoType& type);
   
   //! Constructor for creation of Scans without openDirectory
   BasicScan(double * rPos, double * rPosTheta, std::vector<double*> points);
@@ -67,7 +67,7 @@ private:
   bool m_filter_range_set, m_filter_height_set, m_filter_custom_set, m_range_mutation_set;
   string customFilterStr;
 
-  std::map<std::string, std::pair<unsigned char*, unsigned int> > m_data;
+  std::map<std::string, std::pair<unsigned char*, size_t> > m_data;
 
   std::vector<Frame> m_frames;
